# Project-Server/openapi/model/organizerDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EventDto } from './eventDto';


export interface OrganizerDto { 
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    role: string;
    events?: Array<EventDto>;
    id?: number;
}